requirements.txt
discord.py
requests
python-dotenv
pillow